requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.1.13
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
idna==3.7
migrate==0.3.8
mysqlclient==2.2.4
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pymongo==3.12.3
pytz==2023.4
requests==2.31.0
sqlparse==0.2.4
typing_extensions==4.11.0
urllib3==2.2.1
